+++ Assets/Scripts/UI/GameUI.cs
@@ -74,19 +74,19 @@ public class GameUI : MonoBehaviour {
 	}
 
 	public void updatePlayerHealth(int current_health, int max_health) {
-		health_bar.fillAmount = current_health / max_health;
+		health_bar.fillAmount = (float)current_health / max_health;
 		// TODO: Add numbers text after bar? In bar?
 		player_stats_ui_script.updatePlayerHealth(current_health, max_health);
 	}
 
 	public void updatePlayerStamina(int current_stamina, int max_stamina) {
-		stamina_bar.fillAmount = current_stamina / max_stamina;
+		stamina_bar.fillAmount = (float)current_stamina / max_stamina;
 		// TODO: Add numbers text after bar? In bar?
 		player_stats_ui_script.updatePlayerStamina(current_stamina, max_stamina);
 	}
 
 	public void updatePlayerExperience(int level, int current_experience, int next_level) {
-		experience_bar.fillAmount = current_experience / next_level;
+		experience_bar.fillAmount = (float)current_experience / next_level;
 		experience_text.text = current_experience + "/" + next_level;
 		player_level_text.text = "Level " + level;
